# openapi.yaml
openapi: 3.1.0
info:
  title: assessment-service
  version: v0.3.1
paths:
...
  /v1/admin/fixtures/cleanup:
    post:
      tags:
      - assessment
      - assessment
      summary: Cleanup Fixtures
      operationId: postAdminFixturesCleanup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FixtureCleanupRequest'
            example:
              dry_run: true
              subject_code_patterns:
              - string
              topic_code_patterns:
              - string
              node_id_patterns:
              - string
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureCleanupResponse'
              example:
                status: planned
                dry_run: true
                filters:
                  subject_code_patterns:
                  - string
                  topic_code_patterns:
                  - string
                  node_id_patterns:
                  - string
                matched:
                  subjects: 1
                  topics: 1
                  micro_skills: 1
                  tests: 1
                  questions: 1
                  assignments: 1
                  attempts: 1
                  answers: 1
                deleted:
                  subjects: 1
                  topics: 1
                  micro_skills: 1
                  tests: 1
                  questions: 1
                  assignments: 1
                  attempts: 1
                  answers: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
  /v1/admin/diagnostics/children/{child_id}:
    get:
      tags:
      - assessment
      - assessment
      summary: Get Child Diagnostics
      operationId: getAdminDiagnosticsChildrenByChildId
      parameters:
      - name: child_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Child Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildDiagnosticsResponse'
              example:
                child_id: 00000000-0000-4000-8000-000000000000
                assignments_total: 1
                attempts_total: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/results/children/{child_id}:
    get:
      tags:
      - assessment
      - assessment
      summary: Get Child Results
      operationId: getAdminResultsChildrenByChildId
      parameters:
      - name: child_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Child Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildResultsResponse'
              example:
                child_id: 00000000-0000-4000-8000-000000000000
                summary:
                  attempts_total: 1
                  submitted_attempts_total: 1
                  started_attempts_total: 1
                  attempts_with_diagnostics_total: 1
                  answers_total: 1
                  expected_answers_total: 1
                  correct_answers_total: 1
                  accuracy_percent: 1
                  time_spent_ms_total: 1
                  avg_time_per_answer_ms: 1
                attempts:
                - attempt_id: string
                  assignment_id: string
                  status: string
                  score: 1
                  started_at: '2026-03-13T00:00:00Z'
                  submitted_at: string
                  duration_seconds: string
                  answers_total: 1
                  expected_answers_total: 1
                  unanswered_answers_total: 1
                  correct_answers: 1
                  accuracy_percent: 1
                  time_spent_ms_total: 1
                  avg_time_per_answer_ms: string
                  has_resolved_diagnostics: true
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/results/children/{child_id}/skills:
    get:
      tags:
      - assessment
      - assessment
      summary: Get Child Skill Results
      operationId: getAdminResultsChildrenSkillsByChildId
      parameters:
      - name: child_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Child Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildSkillResultsResponse'
              example:
                child_id: 00000000-0000-4000-8000-000000000000
                summary:
                  total_skills: 1
                  high_gap_total: 1
                  medium_gap_total: 1
                  low_gap_total: 1
                  insufficient_data_total: 1
                skills:
                - node_id: string
                  topic_code: string
                  skill_name: string
                  attempted_questions: 1
                  correct_answers: 1
                  accuracy_percent: 1
                  avg_time_per_answer_ms: string
                  wilson_low: 1
                  wilson_high: 1
                  gap_level: insufficient_data
                  recommendation: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ChildDiagnosticsResponse:
      properties:
        child_id:
          type: string
          format: uuid
          title: Child Id
        assignments_total:
          type: integer
          title: Assignments Total
        attempts_total:
          type: integer
          title: Attempts Total
      type: object
      required:
      - child_id
      - assignments_total
      - attempts_total
      title: ChildDiagnosticsResponse
    ChildResultsAttemptResponse:
      properties:
        attempt_id:
          type: string
          title: Attempt Id
        assignment_id:
          type: string
          title: Assignment Id
        status:
          type: string
          title: Status
        score:
          type: integer
          title: Score
        started_at:
          type: string
          format: date-time
          title: Started At
        submitted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Submitted At
        duration_seconds:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Seconds
        answers_total:
          type: integer
          title: Answers Total
        expected_answers_total:
          type: integer
          title: Expected Answers Total
        unanswered_answers_total:
          type: integer
          title: Unanswered Answers Total
        correct_answers:
          type: integer
          title: Correct Answers
        accuracy_percent:
          type: number
          title: Accuracy Percent
        time_spent_ms_total:
          type: integer
          title: Time Spent Ms Total
        avg_time_per_answer_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Avg Time Per Answer Ms
        has_resolved_diagnostics:
          type: boolean
          title: Has Resolved Diagnostics
        resolved_diagnostic_tags:
          items:
            $ref: '#/components/schemas/ChildResultsDiagnosticTagCountResponse'
          type: array
          title: Resolved Diagnostic Tags
      type: object
      required:
      - attempt_id
      - assignment_id
      - status
      - score
      - started_at
      - submitted_at
      - duration_seconds
      - answers_total
      - expected_answers_total
      - unanswered_answers_total
      - correct_answers
      - accuracy_percent
      - time_spent_ms_total
      - avg_time_per_answer_ms
      - has_resolved_diagnostics
      title: ChildResultsAttemptResponse
    ChildResultsDiagnosticTagCountResponse:
      properties:
        tag:
          $ref: '#/components/schemas/DiagnosticTag'
        count:
          type: integer
          title: Count
      type: object
      required:
      - tag
      - count
      title: ChildResultsDiagnosticTagCountResponse
    ChildResultsResponse:
      properties:
        child_id:
          type: string
          format: uuid
          title: Child Id
        summary:
          $ref: '#/components/schemas/ChildResultsSummaryResponse'
        attempts:
          items:
            $ref: '#/components/schemas/ChildResultsAttemptResponse'
          type: array
          title: Attempts
      type: object
      required:
      - child_id
      - summary
      title: ChildResultsResponse
    ChildResultsSummaryResponse:
      properties:
        attempts_total:
          type: integer
          title: Attempts Total
        submitted_attempts_total:
          type: integer
          title: Submitted Attempts Total
        started_attempts_total:
          type: integer
          title: Started Attempts Total
        attempts_with_diagnostics_total:
          type: integer
          title: Attempts With Diagnostics Total
        answers_total:
          type: integer
          title: Answers Total
        expected_answers_total:
          type: integer
          title: Expected Answers Total
        correct_answers_total:
          type: integer
          title: Correct Answers Total
        accuracy_percent:
          type: number
          title: Accuracy Percent
        time_spent_ms_total:
          type: integer
          title: Time Spent Ms Total
        avg_time_per_answer_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Avg Time Per Answer Ms
        resolved_diagnostic_tags:
          items:
            $ref: '#/components/schemas/ChildResultsDiagnosticTagCountResponse'
          type: array
          title: Resolved Diagnostic Tags
      type: object
      required:
      - attempts_total
      - submitted_attempts_total
      - started_attempts_total
      - attempts_with_diagnostics_total
      - answers_total
      - expected_answers_total
      - correct_answers_total
      - accuracy_percent
      - time_spent_ms_total
      - avg_time_per_answer_ms
      title: ChildResultsSummaryResponse
    ChildSkillResultResponse:
      properties:
        node_id:
          type: string
          title: Node Id
        topic_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic Code
        skill_name:
          type: string
          title: Skill Name
        attempted_questions:
          type: integer
          title: Attempted Questions
        correct_answers:
          type: integer
          title: Correct Answers
        accuracy_percent:
          type: number
          title: Accuracy Percent
        avg_time_per_answer_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Avg Time Per Answer Ms
        wilson_low:
          type: number
          title: Wilson Low
        wilson_high:
          type: number
          title: Wilson High
        gap_level:
          type: string
          enum:
          - insufficient_data
          - high
          - medium
          - low
          title: Gap Level
        resolved_diagnostic_tags:
          items:
            $ref: '#/components/schemas/ChildResultsDiagnosticTagCountResponse'
          type: array
          title: Resolved Diagnostic Tags
        recommendation:
          type: string
          title: Recommendation
      type: object
      required:
      - node_id
      - topic_code
      - skill_name
      - attempted_questions
      - correct_answers
      - accuracy_percent
      - avg_time_per_answer_ms
      - wilson_low
      - wilson_high
      - gap_level
      - recommendation
      title: ChildSkillResultResponse
    ChildSkillResultsResponse:
      properties:
        child_id:
          type: string
          format: uuid
          title: Child Id
        summary:
          $ref: '#/components/schemas/ChildSkillResultsSummaryResponse'
        skills:
          items:
            $ref: '#/components/schemas/ChildSkillResultResponse'
          type: array
          title: Skills
      type: object
      required:
      - child_id
      - summary
      title: ChildSkillResultsResponse
    ChildSkillResultsSummaryResponse:
      properties:
        total_skills:
          type: integer
          title: Total Skills
        high_gap_total:
          type: integer
          title: High Gap Total
        medium_gap_total:
          type: integer
          title: Medium Gap Total
        low_gap_total:
          type: integer
          title: Low Gap Total
        insufficient_data_total:
          type: integer
          title: Insufficient Data Total
      type: object
      required:
      - total_skills
      - high_gap_total
      - medium_gap_total
      - low_gap_total
      - insufficient_data_total
      title: ChildSkillResultsSummaryResponse
...
    DiagnosticTag:
      type: string
      enum:
      - inattention
      - misread_condition
      - calc_error
      - concept_gap
      - guessing
      - other
      title: DiagnosticTag
    FixtureCleanupCountsResponse:
      properties:
        subjects:
          type: integer
          title: Subjects
        topics:
          type: integer
          title: Topics
        micro_skills:
          type: integer
          title: Micro Skills
        tests:
          type: integer
          title: Tests
        questions:
          type: integer
          title: Questions
        assignments:
          type: integer
          title: Assignments
        attempts:
          type: integer
          title: Attempts
        answers:
          type: integer
          title: Answers
      type: object
      required:
      - subjects
      - topics
      - micro_skills
      - tests
      - questions
      - assignments
      - attempts
      - answers
      title: FixtureCleanupCountsResponse
    FixtureCleanupFiltersResponse:
      properties:
        subject_code_patterns:
          items:
            type: string
          type: array
          title: Subject Code Patterns
        topic_code_patterns:
          items:
            type: string
          type: array
          title: Topic Code Patterns
        node_id_patterns:
          items:
            type: string
          type: array
          title: Node Id Patterns
      type: object
      required:
      - subject_code_patterns
      - topic_code_patterns
      - node_id_patterns
      title: FixtureCleanupFiltersResponse
    FixtureCleanupRequest:
      properties:
        dry_run:
          type: boolean
          title: Dry Run
          default: true
        subject_code_patterns:
          items:
            type: string
          type: array
          title: Subject Code Patterns
        topic_code_patterns:
          items:
            type: string
          type: array
          title: Topic Code Patterns
        node_id_patterns:
          items:
            type: string
          type: array
          title: Node Id Patterns
      type: object
      title: FixtureCleanupRequest
    FixtureCleanupResponse:
      properties:
        status:
          type: string
          enum:
          - planned
          - completed
          title: Status
        dry_run:
          type: boolean
          title: Dry Run
        filters:
          $ref: '#/components/schemas/FixtureCleanupFiltersResponse'
        matched:
          $ref: '#/components/schemas/FixtureCleanupCountsResponse'
        deleted:
          $ref: '#/components/schemas/FixtureCleanupCountsResponse'
      type: object
      required:
      - status
      - dry_run
      - filters
      - matched
      - deleted
      title: FixtureCleanupResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ProblemDetails:
      type: object
      title: ProblemDetails
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        code:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/authentication
            title: Unauthorized
            status: 401
            detail: Bearer access token is required
            code: UNAUTHORIZED
            request_id: 00000000-0000-4000-8000-000000000000
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/access-denied
            title: Forbidden
            status: 403
            detail: Access denied
            code: FORBIDDEN
            request_id: 00000000-0000-4000-8000-000000000000
    TooManyRequests:
      description: Too Many Requests
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/too-many-requests
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded
            code: RATE_LIMITED
            request_id: 00000000-0000-4000-8000-000000000000
servers:
- url: http://localhost:8003
  description: Local
- url: https://api.monitoring.example
  description: Production
security:
- bearerAuth: []